remove mongo repository, merge dbHelper into mongo.js
