+++ src/data/mongo.js
@@ -6,6 +6,27 @@ const client = new MongoClient(uri);
 
 let db = null;
 
+async function getDatabaseCollection(collectionName) {
+   try {
+      if (!db) {
+         db = await connectToDatabase();
+      }
+      return db.collection(collectionName);
+   } catch (err) {
+      console.error("Error getting database collection:", err);
+      throw err;
+   }
+}
+
+async function executeWithCatch(context, operation) {
+   try {
+      return await operation();
+   } catch (err) {
+      console.error(`Error in ${context}:`, err);
+      throw err;
+   }
+}
+
 async function connectToDatabase() {
    if (db) return db; // Return the existing connection if already connected
 
@@ -20,4 +41,7 @@ async function connectToDatabase() {
    }
 }
 
-module.exports = { connectToDatabase };
+module.exports = {
+   getDatabaseCollection,
+   executeWithCatch,
+};
