+++ src/service/storyInput.js
@@ -1,6 +1,6 @@
-const ConfigRepository = require("@data/mongo/config.js");
-const StoryRepository = require("@data/mongo/story.js");
-const StoryInputRepository = require("@data/mongo/storyInput.js");
+const ConfigRepository = require("@data/config.js");
+const StoryRepository = require("@data/story.js");
+const StoryInputRepository = require("@data/storyInput.js");
 
 class StoryInputService {
    async insertStoryInput(message) {
